+++ packages/core/src/index.ts
@@ -7,11 +7,8 @@ export {
   type AMAContextValue,
 } from './components/AMAProvider';
 
-export { Pressable, type PressableProps } from './components/Pressable';
-
 // Hooks
 export { useButtonChecks } from './hooks/useButtonChecks';
 export { useChecks } from './hooks/useChecks';
 export { useFocus } from './hooks/useFocus';
-export { usePressable, type UsePressable } from './hooks/usePressable';
 export { useTimedAction } from './hooks/useTimedAction';
