+++ packages/react-native/src/components/TouchableOpacity.test.tsx
@@ -1,4 +1,4 @@
-import * as UsePressable from '@react-native-ama/core/src/hooks/usePressable';
+import * as UsePressable from '@react-native-ama/react-native/src/hooks/usePressable';
 import { render } from '@testing-library/react-native';
 import * as React from 'react';
