+++ packages/react-native/src/hooks/useExpandable.test.ts
@@ -1,7 +1,7 @@
-import { type PressableProps } from '@react-native-ama/core';
 import * as UseChecks from '@react-native-ama/core/src/hooks/useChecks';
 import { renderHook } from '@testing-library/react-native';
 
+import { type PressableProps } from '../components/Pressable';
 import { useExpandable } from './useExpandable';
 
 beforeEach(() => {
